update(globalTheme): 클로벌 테마 타입스크립트 적용
