--- src/scripts/Components/GlobalTheme/GlobalTheme.js
@@ -1,15 +1,0 @@
-import { createMuiTheme } from "@material-ui/core/styles";
-
-const GlobalTheme = createMuiTheme({
-  palette: {
-    jettBlack: {
-      100: "#45433e",
-      200: "#66635c",
-      300: "#99958b",
-      400: "#ccc6b9",
-      500: "#fff8e7",
-    },
-  },
-});
-
-export default GlobalTheme;
