+++ src/scripts/Components/GlobalTheme/GlobalTheme.tsx
@@ -1,0 +1,13 @@
+import { createMuiTheme } from "@material-ui/core/styles";
+
+const GlobalTheme = createMuiTheme({
+  jettBlack: {
+    100: "#45433e",
+    200: "#66635c",
+    300: "#99958b",
+    400: "#ccc6b9",
+    500: "#fff8e7",
+  },
+});
+
+export default GlobalTheme;
